store stroked symbol coords as double not int to avoid some rounding wiggle
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@57021 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/symbol.h
@@ -51,7 +51,7 @@ typedef struct
     int count, alloc;		/* number of elements */
     SYMBEL **elem;		/* array of elements */
     int scount, salloc;		/* number of points in stroked version */
-    int *sx, *sy;		/* coordinates in stroked version */
+    double *sx, *sy;		/* coordinates in stroked version */
 } SYMBCHAIN;
 
 /* part */
